Mixer: Handle initial negative throttle correctly
When motor_bidir_reverse is true, most of the time the input will
be initially negative, so instead of biasing the check towards the
positive throttle we have to properly handle the deadband transition
and clipping.

Also, remove branching from the last motor assignment loop by
introducing 2 extra variables for clipped THR boundaries instead
of overwriting the non clipped ones.